% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaSampling.R
\name{generateBetaRandomVariable}
\alias{generateBetaRandomVariable}
\title{Generate beta distribution samples using R}
\usage{
generateBetaRandomVariable(a, b)
}
\arguments{
\item{a}{the first positive shape parameter of beta distribution}

\item{b}{the second positive shape parameter of beta distribution}
}
\value{
a random sample of size \code{n}
}
\description{
Generate beta distribution samples using R
}
\examples{
\dontrun{
rnR <- generateBetaRandomVariable(2,5)
plot(rnR,type='l')
}
}
