% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCauthyMetropolisHastings.R
\name{generateCauthyMetropolisHastings}
\alias{generateCauthyMetropolisHastings}
\title{Generate cauthy distribution with Metropolis Hastings method}
\usage{
generateCauthyMetropolisHastings(n, sigma, x0, N)
}
\arguments{
\item{n}{degree of freedom of t distribution}

\item{sigma}{standard variance of proposal distribution}

\item{x0}{initial value}

\item{N}{size of random numbers required}
}
\value{
a random samples of cauthy distribution
}
\description{
Generate cauthy distribution with Metropolis Hastings method
}
\examples{
\dontrun{
n <- 1
x0 <- 10
sigma <-3
N <- 5000
rnR <- generateCauthyMetropolisHastings(n,sigma,x0,N)
print(rnR)
}
}
